Bug fixes for nanowm, client exit cleanup, TinyWidgets support and patches

+++ src/mwin/winres.c
@@ -390,7 +390,7 @@ resGetDlgItemTemplExtra(PMWDLGITEMTEMPLATE pItem, PMWDLGITEMTEMPLEXTRA pItemExtr
 
 	pw += 1 + ((*pw) >> 1);
 	//  dword align
-	if ((((unsigned long) pw) & 2) != 0)
+	if ((((DWORD)pw) & 2) != 0)
 		pw++;
 	return (PMWDLGITEMTEMPLATE) pw;
 }
@@ -492,7 +492,7 @@ resFirstDlgItem(PMWDLGTEMPLATE pDlg)
 		RES_SKIP_WSTRING(pw);	// skip font name
 	}
 	//  dword align
-	if ((((uint32_t) pw) & 2) != 0)
+	if ((((DWORD)pw) & 2) != 0)
 		pw++;
 	return (PMWDLGITEMTEMPLATE) pw;
 }
@@ -510,7 +510,7 @@ resNextDlgItem(PMWDLGITEMTEMPLATE pItem)
 	RES_SKIP_WSTRING(pw);
 	pw += 1 + ((*pw) >> 1);
 	//  dword align
-	if ((((uint32_t) pw) & 2) != 0)
+	if ((((DWORD)pw) & 2) != 0)
 		pw++;
 	return (PMWDLGITEMTEMPLATE) pw;
 }
